Extend OPUS interface (1)
Move file interface to separate file
