+++ tum_esm_utils/opus/utils.py
@@ -1,4 +1,6 @@
-"""Credits to Friedrich Klappenbach (ge79wul@mytum.de) for decoding the OPUS file format."""
+"""Utilities for interacting with OPUS files.
+
+Credits to Friedrich Klappenbach (ge79wul@mytum.de) for decoding the OPUS file format."""
 
 from __future__ import annotations
 from typing import Optional, TypeVar
